// apps/landing/src/services/models/ServiceViewStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { HttpStatusCode } from './HttpStatusCode';
import type { ServiceView } from './ServiceView';

export type ServiceViewStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: ServiceView;
    statusCode?: HttpStatusCode;
    errors?: any;
};